created doc on adjusting video frames per second
walkthrough how to increase the frames per second for the video camera, but will only work if the camera supports the number of FPS.

Current problem facing is finding the resolution or window size and FPS supported FPS with the video camera device.

I am able to see the camera connected using lsusb, but when I try to use video 4 linux command, it doesn't pick up the camera.